+++ src/components/EventScreens/EventMenuStack/EventInfoForm.js
@@ -1,16 +1,15 @@
 import React, { useEffect, useState } from 'react'
 import t from 'tcomb-form-native'
 import RNDateTimePicker from '@react-native-community/datetimepicker'
-import { formatDate } from '../../utils/dates'
+import { formatDate } from '../../../utils/dates'
 import { Button } from 'react-native-elements'
-import { View, Dimensions } from 'react-native'
-import { paddingSides } from '../../styles/global'
-import { startEvent, setOngoingEvent } from '../../reducers/eventReducer'
+import { View } from 'react-native'
+import { paddingSides } from '../../../styles/global'
+import { startEvent, setOngoingEvent } from '../../../reducers/eventReducer'
 import { connect } from 'react-redux'
-import { inOneHour } from '../../utils/dates'
-import BackgroundImage from '../common/BackgroundImage'
-import AppButton from '../common/AppButton'
-import targetService from '../../services/targets'
+import { inOneHour } from '../../../utils/dates'
+import AppButton from '../../common/AppButton'
+import targetService from '../../../services/targets'
 import { useIsFocused } from '@react-navigation/native'
 import _ from 'lodash'
